Add gesture event regions to EventRegion https://bugs.webkit.org/show_bug.cgi?id=302281 rdar://164343188
Reviewed by Abrar Rahman Protyasha and Wenson Hsieh.

Add gesture event regions to EventRegion to ensure that gesture
events are correctly dispatched when touch event regions are
enabled.

Reorder members of StyleRareInheritedData to reduce padding
and offset growth.

* Source/WebCore/page/scrolling/ScrollingTree.h:
* Source/WebCore/platform/EventTrackingRegions.cpp:
(WebCore::EventTrackingRegions::eventName):
(WebCore::EventTrackingRegions::eventNameAtomString):
* Source/WebCore/platform/EventTrackingRegions.h:
* Source/WebCore/rendering/EventRegion.cpp:
(WebCore::eventTypeForEventListenerType):
* Source/WebCore/rendering/RenderLayerBacking.cpp:
(WebCore::patternForEventListenerRegionType):
* Source/WebCore/rendering/style/RenderStyle.h:
* Source/WebCore/rendering/style/RenderStyleConstants.cpp:
(WebCore::operator<<):
* Source/WebCore/rendering/style/RenderStyleConstants.h:
* Source/WebCore/rendering/style/StyleRareInheritedData.cpp:
(WebCore::StyleRareInheritedData::StyleRareInheritedData):
(WebCore::usedAppleVisualEffectForSubtree):
* Source/WebCore/rendering/style/StyleRareInheritedData.h:
* Source/WebCore/style/StyleAdjuster.cpp:
(WebCore::Style::Adjuster::computeEventListenerRegionTypes):
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:
* Source/WebKit/UIProcess/RemoteLayerTree/cocoa/RemoteScrollingTreeCocoa.h:
* Source/WebKit/UIProcess/RemoteLayerTree/ios/RemoteLayerTreeViews.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::updateTouchEventTracking):

Canonical link: https://commits.webkit.org/302954@main

+++ Source/WebCore/rendering/style/RenderStyle.h
@@ -115,7 +115,7 @@ enum class CursorVisibility : bool;
 enum class DisplayType : uint8_t;
 enum class DominantBaseline : uint8_t;
 enum class EmptyCell : bool;
-enum class EventListenerRegionType : uint32_t;
+enum class EventListenerRegionType : uint64_t;
 enum class FieldSizing : bool;
 enum class FillAttachment : uint8_t;
 enum class FillBox : uint8_t;
